refactor: move util function and tests
