+++ src/cabinUt/cabinUt.utils.test.ts
@@ -1,18 +1,6 @@
 import * as assert from 'node:assert';
 import { describe, it } from 'node:test';
-import { dateIsValid, getVisbookId, isReservationsUrl } from './cabinUt.utils';
-
-describe('Validating input as date string yyyy-mm-dd', () => {
-  it('returns true for valid date string', () => {
-    assert.deepEqual(dateIsValid('2022-10-16'), true);
-  });
-  it('returns false for invalid date string', () => {
-    assert.deepEqual(dateIsValid('2022-10-66'), false);
-  });
-  it('returns false for any string', () => {
-    assert.deepEqual(dateIsValid('hello'), false);
-  });
-});
+import { getVisbookId, isReservationsUrl } from './cabinUt.utils';
 
 describe('Extract visbook cabin id from booking URL', () => {
   it('finds id in reservations url', () => {
